E2E: Check SNS balance after getTokens
