+++ frontend/src/tests/page-objects/Accounts.page-object.ts
@@ -1,6 +1,7 @@
-import { BasePageObject } from "$tests/page-objects/base.page-object";
 import { NnsAccountsPo } from "$tests/page-objects/NnsAccounts.page-object";
 import { NnsAccountsFooterPo } from "$tests/page-objects/NnsAccountsFooter.page-object";
+import { SnsAccountsPo } from "$tests/page-objects/SnsAccounts.page-object";
+import { BasePageObject } from "$tests/page-objects/base.page-object";
 import type { PageObjectElement } from "$tests/types/page-object.types";
 
 export class AccountsPo extends BasePageObject {
@@ -12,6 +13,10 @@ export class AccountsPo extends BasePageObject {
 
   getNnsAccountsPo(): NnsAccountsPo {
     return NnsAccountsPo.under(this.root);
+  }
+
+  getSnsAccountsPo(): SnsAccountsPo {
+    return SnsAccountsPo.under(this.root);
   }
 
   getNnsAccountsFooterPo(): NnsAccountsFooterPo {
